+++ src/diffusers/pipelines/ltx2/pipeline_output.py
@@ -30,16 +30,16 @@ class LTX2DFRPipelineOutput(LTX2PipelineOutput):
 
     Args:
         frames (`torch.Tensor`, `np.ndarray`, or list[list[PIL.Image.Image]]):
-            Denoised video. Latent output is the untrimmed canvas, shape
-            `(batch_size, num_channels, latent_frames, latent_height, latent_width)`.
+            Denoised video. Latent output is the untrimmed canvas, shape `(batch_size, num_channels, latent_frames,
+            latent_height, latent_width)`.
         audio (`torch.Tensor`, `np.ndarray`):
             Accompanying audio latents or waveform.
         keyframes (`torch.Tensor`, *optional*):
             Generated or carried keyframe latents of shape `(batch_size, num_channels, num_keyframes, latent_height,
             latent_width)`. `None` when the pass did not produce slots (e.g. a tiled epilogue).
         keyframe_positions (`list[int]`, *optional*):
-            Pixel-frame index of each keyframe on this pass's canvas. After a temporal round these cannot be
-            re-derived from the original `num_frames` and must be passed into the next stage.
+            Pixel-frame index of each keyframe on this pass's canvas. After a temporal round these cannot be re-derived
+            from the original `num_frames` and must be passed into the next stage.
     """
 
     keyframes: torch.Tensor | None = None
